feat: add setOperator function
Added setOperator function that checks if the current expression
already has an operator or not, if it does, the evaluate function is
called. Otherwise, the current expression's operator is set to the
passed in value.

This is so that we can store the operator and the first number of the
expression before getting the second number that will be passed into
the operate function.