refactor: split resetOutput function into delOne
Refactored the resetOutput function to only completely clear the
text content of the output display, then created a new function,
delOne, which deletes only 1 character from the output.

This is because the app now takes keyboard input, and the resetOutput
function was not going to work the way it was with keyboard inputs.
This also cleans up the code a little, more readability, and
there's less lines.